Add ability to inactive Websites
We needed a way to inactivate websites for when we need to do
testing/blue/green deploys.

Fixes #145

When a website is inactivated:
- The PingerJob will re-enqueue in 1 minute
- This was the easy way out rather than getting into callback-hell.

Make a website inactive by editing it in the views.

Adds support to the Website Creation API/Form to create inactive
websites.

Currently there is no support yet for updating existing websites via the
API.