Remove default_scope for reordering records
Using a default scope to reorder records is not always the best
idea, since it's easy to introduce bugs when using `.first` and
`.last` methods.

This change fixes an issue when trying to fetch the latest ping
for a website, using `.last`, for the retry logic, we were actually
fetching the first ping.